# specs/001-ocr-translation-viewer/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OCR Translation Comparison Viewer API
  description: |
    REST API for loading document sets, navigating pages, and managing viewer state.
    Implements server-side processing for PDF rendering and markdown parsing.
  version: 1.0.0
  contact:
    name: AI-OCR Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://ocr-viewer.example.com/api
    description: Production server
...
paths:
...
  /documents/{documentId}/validate:
    post:
      tags:
        - documents
      summary: Validate document structure
      description: |
        Checks folder structure, file naming conventions, and page counts.
        Returns validation errors if any issues found.
      operationId: validateDocument
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean
                    example: true
                  errors:
                    type: array
                    items:
                      type: string
                    example: []
                  warnings:
                    type: array
                    items:
                      type: string
                    example: ['Page 47 missing in raw.en-US folder']
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /documents/{documentId}/pages/{pageNumber}/pdf:
    get:
      tags:
        - pages
      summary: Get PDF page content
      description: |
        Renders a specific PDF page and returns canvas data or image.
        Supports lazy loading for performance.
      operationId: getPdfPage
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
        - name: pageNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            example: 5
        - name: scale
          in: query
          schema:
            type: number
            minimum: 0.5
            maximum: 3.0
            default: 1.0
          description: Render scale factor for high-DPI displays
      responses:
        '200':
          description: PDF page rendered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PDFPageContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: PDF file exceeds maximum size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PDFPageContent:
      type: object
      required:
        - pageNumber
        - width
        - height
        - imageDataUrl
      properties:
        pageNumber:
          type: integer
          minimum: 1
          example: 5
        width:
          type: number
          example: 612
        height:
          type: number
          example: 792
        imageDataUrl:
          type: string
          format: uri
          description: Base64-encoded PNG data URL
          example: 'data:image/png;base64,iVBORw0KGgoAAAANS...'
        textContent:
          type: string
          description: Extracted text (optional, for future search)
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - DOCUMENT_NOT_FOUND
            - INVALID_PAGE_NUMBER
            - INVALID_DOCUMENT_ID
            - INVALID_LANGUAGE_CODE
            - PDF_TOO_LARGE
            - CORRUPTED_PDF
            - MISSING_DATA_FOLDER
            - INVALID_FOLDER_STRUCTURE
            - INTERNAL_ERROR
          example: 'DOCUMENT_NOT_FOUND'
        message:
          type: string
          description: Human-readable error message
          example: 'Document with ID "contract-2024" not found'
        details:
          type: object
          description: Additional error context (optional)
          additionalProperties: true
          example:
            maxSizeMB: 50
            actualSizeMB: 75
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'INVALID_PAGE_NUMBER'
            message: 'Page number must be between 1 and 87'
    PayloadTooLarge:
      description: PDF file exceeds maximum size limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'PDF_TOO_LARGE'
            message: 'PDF file exceeds maximum allowed size'
            details:
              maxSizeMB: 50
              actualSizeMB: 75
              configKey: 'MAX_PDF_SIZE_MB'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'DOCUMENT_NOT_FOUND'
            message: 'Document with ID "contract-2024" not found'
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 'INTERNAL_ERROR'
            message: 'Failed to read data folder'
            details:
              path: '/data'
              reason: 'Permission denied'
  securitySchemes:
    # No authentication required for initial version
    # Future: Add API key or OAuth2 if needed
    {}